enable xray on core scheduler and move tests to test directory
